Added Caching
Putting a cache of 1 item has more than
halved the number of reads we need to do.
Reading 1000 entries now only triggers
411 calls to the OS read, vs 1166 before.

This is unbelieveable but makes sense
given the fact that most recent data
floats to the beginning of the SSTable
list and likely
has the most relevant pages.

In general this won't work but it was
a fun experiment. Might be worth having
in some use cases.

This might be even less useful with
compression feature added, but
I still suspect it'll be useful regardless.